docs(preamble): :memo: add yml header with tbl-colwidths (#49)
## Description

This aligns the column width across schedule tables. 

I found out that the difference in col widths (Day 1 and 2 were the same
and Day 3 slightly different) were because the first two days having longer
texts (the social activity with networking). We can remedy this by
fixing the `tbl-colwidths` in the YAML header.

## Reviewer Focus

This PR needs a quick review. 

## Checklist

- [ ] Ran spell check
- [ ] Formatted Markdown
- [X] Rendered website locally