docs(preamble): 📝 update sessions in schedule (#51)
## Description

I have added the schedule based on the discussions in #2 to the
schedule. I thought that it would make things a bit clearer to see the
course/sessions in the schedule, but the exact sessions/schedule is
still up for debate. What do you think?

Closes #6

This PR needs an in-depth review. 

## Checklist

- [X] Ran spell-check
- [X] Formatted Markdown
- [X] Rendered website locally

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>