docs: 🎨 set course title and subtitle (#57)
## Description

This PR adds the course title and subtitle in the `_quarto.yml`. 

Initially, I added the title and subtitle we agreed on in #55 and #40 ,
but I kept brainstorming a bit and found an alternative that I liked
more; it's simpler/shorter and more digestible, in my opinion.

There's just something that I don't quite like about "collaborating as a
team" + we're missing the *s* from the abbreviation. And I liked the
explicit addition of the "CATs" abbreviation in the title too.
With the subtitle "An introduction to ...", I think that we still
incorporate the "you don't have to be collaborative team just yet, but
wanting to learn it is enough"-part that we talked about earlier.

Buuutttt, if you don't like it, I can just revert it 🐱 🐈 

Closes #55 
Closes #40
Related to #20 

## Reviewer Focus

This PR needs an in-depth review. 

## Checklist

- [X] Rendered website locally

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>