docs: 📝 update text by simplifying a bit (#60)
## Description

This PR completes some minor text edits in the `syllabus.qmd`. 

Mainly, I have tried to remove words and sentences that are (/that I
deem) unnecessary to try to keep things a bit simpler and more to the
point :)

I also fixed a layout issue with the nested lists. 

Closes #14 (this issue is the reason why I started going through this
file, but it was almost filled out already)

## Reviewer Focus

This PR needs an in-depth review. 

## Checklist

- [X] Ran spell-check
- [X] Formatted Markdown
- [X] Rendered website locally

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>