chore: :truck: use symlink of extension file (#94)
## Description

This change is done to simplify the build of the slides, by using a
symlink of the extensions folder into the slides folder. The reason
being is that when rendering a Quarto doc on its own (which we need to
do for the slides for book projects since slides aren't usable in
books), Quarto needs the extension in the same location as the doc.
Hence, using the symlink.

## Reviewer Focus

This PR needs a quick review.